Inclusao de HATEOAS

+++ src/main/java/com/alesaudate/samples/springjersey/services/BaseService.java
@@ -10,19 +10,21 @@ import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
+import javax.ws.rs.QueryParam;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.Response.Status;
 import javax.ws.rs.core.UriBuilder;
 
 import org.springframework.security.access.prepost.PreAuthorize;
 
-import com.alesaudate.samples.springjersey.entities.BaseEntity;
+import com.alesaudate.samples.springjersey.entities.HATEOASEntity;
 import com.alesaudate.samples.springjersey.persistence.GenericDao;
 
 
 @Consumes(MediaType.APPLICATION_XML) //Could easily be changed to JSON
 @Produces(MediaType.APPLICATION_XML)
-public abstract class BaseService<T extends BaseEntity> {
+public abstract class BaseService<T extends HATEOASEntity> {
 	
 	
 	
@@ -40,6 +42,10 @@ public abstract class BaseService<T extends BaseEntity> {
 	@POST
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
 	public Response create(T entity) {
+		if (entity == null) {
+			return Response.status(Status.BAD_REQUEST).build();
+		}
+		entity.setId(null);
 		dao.create(entity);
 		return Response.created(UriBuilder.fromPath("/" + entity.getId()).build()).build();
 	}
@@ -47,43 +53,62 @@ public abstract class BaseService<T extends BaseEntity> {
 	@GET
 	@Path("/{id}")
 	@PreAuthorize("hasRole('ROLE_USER')")
-	public T retrieve(@PathParam("id") Long id) {
-		return dao.retrieve(id);
+	public Response retrieve(@PathParam("id") Long id) {
+		T entity = dao.retrieve(id);
+		if (entity != null) {
+			entity.createSelfLink();
+			return Response.ok(entity).build();
+		}
+		
+		return Response.status(Status.NOT_FOUND).build();
 	}
 	
 	
 	@GET
 	@com.alesaudate.samples.springjersey.services.Collection
 	@PreAuthorize("hasRole('ROLE_USER')")
-	public Collection<T> retrieve() {
-		return dao.retrieve();
-	}
-	
+	public Collection<T> retrieve(@QueryParam("page") int page) {
+		Collection<T> entities = dao.retrieve(page);
+		for (T entity : entities) {
+			entity.createSelfLink();
+		}
+		return entities;
+	}	
 	
 	@PUT
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
-	public void update(T entity) {
+	public Response update(T entity) {
+		if (entity == null) {
+			return Response.status(Status.BAD_REQUEST).build();
+		}
 		dao.update(entity);
+		return Response.ok(entity).build();
 	}
 	
 	
 	@DELETE
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
-	public void delete(T entity) {
+	public Response delete(T entity) {
+		if (entity == null) {
+			return Response.status(Status.BAD_REQUEST).build();
+		}
 		dao.delete(entity);
+		return Response.ok().build();
 	}
 	
 	
+	@SuppressWarnings("unchecked")
 	@DELETE
 	@Path("/{id}")
 	@PreAuthorize("hasRole('ROLE_ADMIN')")
-	public void delete(@PathParam("id")Long id) {
-		T entity = retrieve(id);
+	public Response delete(@PathParam("id")Long id) {
+		T entity = (T)retrieve(id).getEntity();
+		if (entity == null) {
+			return Response.status(Status.NOT_FOUND).build();
+		}
 		delete(entity);
+		return Response.ok().build();
 	}
-	
-	 
-
 	
 	
 }
